02-functions/code/submit/requirements.txt
google-cloud-firestore>=2.13.0
google-cloud-pubsub>=2.18.0
google-cloud-storage>=2.10.0
google-auth>=2.22.0
